+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Fixes for the camera roll order in consignments - orta
 * Adds the new photo and selects it when taking a photo - orta
+* Adds a confirmation screen to the end of the consignment submission process - maxim/orta
 
 ### 1.4.0-beta.12